Restructure Tests folders and split runtime artifact test responsibilities
