+++ UniversalToolchain/Tests/Contracts/CompiledArtifactSessionTests.cs
