+++ UniversalToolchain/Tests/Contracts/CompiledArtifactTests.cs
